Connect Footer component
