+++ app/components/Footer/__tests__/Footer-test.js
@@ -5,7 +5,7 @@ import { expect } from 'chai';
 // see: https://github.com/mochajs/mocha/issues/1847
 const { describe, it } = global;
 
-import Footer from '../presenter';
+import Footer from '../default';
 
 
 describe('<Footer />', () => {
